requirements.txt
ctransformers
sentence-transformers
pinecone-client
langchain-huggingface
pinecone
langchain
flask
pypdf
langchain-pinecone
streamlit
langchain-community
python-dotenv